feat & refactor : CS POST 게시글 단건 조회 및 서비스 레이어 나눔
